Move from util/conv to util/url: GetHash(), AddHash(), hasParameter()
