+++ src/astra/stream.go
@@ -10,7 +10,6 @@ import (
 
 	"m3u_merge_astra/cfg"
 	"m3u_merge_astra/deps"
-	"m3u_merge_astra/util/conv"
 	"m3u_merge_astra/util/copier"
 	"m3u_merge_astra/util/network"
 	"m3u_merge_astra/util/slice"
@@ -89,11 +88,11 @@ func (s Stream) UpdateInput(r deps.Global, newURL string, callback func(string))
 			if updRec.From.MatchString(oldURL) && updRec.To.MatchString(newURL) {
 				// Append old hash to new URL.
 				if cfg.KeepInputHash {
-					oldHash, err := conv.GetHash(oldURL)
+					oldHash, err := urlUtil.GetHash(oldURL)
 					if err != nil {
 						r.Log().Debug(err)
 					}
-					newURL, _, err = conv.AddHash(oldHash, newURL)
+					newURL, _, err = urlUtil.AddHash(oldHash, newURL)
 					if err != nil {
 						r.Log().Debug(err)
 					}
@@ -468,7 +467,7 @@ func (r repo) AddHashes(streams []Stream) (out []Stream) {
 			for _, rule := range r.cfg.Streams.InputToInputHashMap {
 				if rule.By.MatchString(inp) {
 					var err error
-					inp, changed, err = conv.AddHash(rule.Hash, inp)
+					inp, changed, err = urlUtil.AddHash(rule.Hash, inp)
 					if err != nil {
 						r.log.Debug(err)
 					}
@@ -481,7 +480,7 @@ func (r repo) AddHashes(streams []Stream) (out []Stream) {
 			for _, rule := range r.cfg.Streams.NameToInputHashMap {
 				if rule.By.MatchString(s.Name) {
 					var err error
-					inp, changed, err = conv.AddHash(rule.Hash, inp)
+					inp, changed, err = urlUtil.AddHash(rule.Hash, inp)
 					if err != nil {
 						r.log.Debug(err)
 					}
@@ -494,7 +493,7 @@ func (r repo) AddHashes(streams []Stream) (out []Stream) {
 			for _, rule := range r.cfg.Streams.GroupToInputHashMap {
 				if rule.By.MatchString(s.FirstGroup()) {
 					var err error
-					inp, changed, err = conv.AddHash(rule.Hash, inp)
+					inp, changed, err = urlUtil.AddHash(rule.Hash, inp)
 					if err != nil {
 						r.log.Debug(err)
 					}
