+++ src/util/conv/string_test.go
@@ -69,157 +69,10 @@ func TestIsNameSame(t *testing.T) {
 	assert.False(t, IsNameSame(cfg, "name_3_var_2", "Name 3"), msg)
 }
 
-func TestGetHah(t *testing.T) {
-	// Good URL's
-	hash, err := GetHash("http://url")
-	assert.Empty(t, hash, "should return empty string")
-	assert.NoError(t, err, "should not return error")
-
-	hash, err = GetHash("http://url#a")
-	assert.Exactly(t, "a", hash, "should return hash")
-	assert.NoError(t, err, "should not return error")
-
-	hash, err = GetHash("http://url#a&b")
-	assert.Exactly(t, "a&b", hash, "should return hash")
-	assert.NoError(t, err, "should not return error")
-
-	// Bad URL's
-	hash, err = GetHash("http://{bad/url")
-	assert.Empty(t, hash, "should not retun hash from invalid URL")
-	assert.Error(t, err, "should return error")
-
-	hash, err = GetHash("http://{bad/url#a")
-	assert.Empty(t, hash, "should not retun hash from invalid URL")
-	assert.Error(t, err, "should return error")
-
-	hash, err = GetHash("http://{bad/url#a&b")
-	assert.Empty(t, hash, "should not retun hash from invalid URL")
-	assert.Error(t, err, "should return error")
-}
-
-func TestAddHash(t *testing.T) {
-	// Good URL's
-	result, changed, err := AddHash("", "http://url")
-	assert.Exactly(t, "http://url", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("", "http://url#a")
-	assert.Exactly(t, "http://url#a", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("a", "http://url")
-	assert.Exactly(t, "http://url#a", result, "should get new hash")
-	assert.True(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("#a", "http://url#a")
-	assert.Exactly(t, "http://url#a", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("b", "http://url#a")
-	assert.Exactly(t, "http://url#a&b", result, "should merge hashes")
-	assert.True(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("a", "http://url#a&b")
-	assert.Exactly(t, "http://url#a&b", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("c", "http://url#a&b")
-	assert.Exactly(t, "http://url#a&b&c", result, "should merge hashes")
-	assert.True(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("c&d", "http://url#a&b")
-	assert.Exactly(t, "http://url#a&b&c&d", result, "should merge hashes")
-	assert.True(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("#c&d", "http://url#a&b")
-	assert.Exactly(t, "http://url#a&b&c&d", result, "should merge hashes")
-	assert.True(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	// Bad URL's
-	result, changed, err = AddHash("", "http://{bad/url")
-	assert.Exactly(t, "http://{bad/url", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("", "http://{bad/url#a")
-	assert.Exactly(t, "http://{bad/url#a", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.NoError(t, err, "should not return error")
-
-	result, changed, err = AddHash("a", "http://{bad/url")
-	assert.Exactly(t, "http://{bad/url", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.Error(t, err, "should return error")
-
-	result, changed, err = AddHash("#a", "http://{bad/url#a")
-	assert.Exactly(t, "http://{bad/url#a", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.Error(t, err, "should return error")
-
-	result, changed, err = AddHash("b", "http://{bad/url#a")
-	assert.Exactly(t, "http://{bad/url#a", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.Error(t, err, "should return error")
-
-	result, changed, err = AddHash("a", "http://{bad/url#a&b")
-	assert.Exactly(t, "http://{bad/url#a&b", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.Error(t, err, "should return error")
-
-	result, changed, err = AddHash("c", "http://{bad/url#a&b")
-	assert.Exactly(t, "http://{bad/url#a&b", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.Error(t, err, "should return error")
-
-	result, changed, err = AddHash("c&d", "http://{bad/url#a&b")
-	assert.Exactly(t, "http://{bad/url#a&b", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.Error(t, err, "should return error")
-
-	result, changed, err = AddHash("#c&d", "http://{bad/url#a&b")
-	assert.Exactly(t, "http://{bad/url#a&b", result, "should stay unmodified")
-	assert.False(t, changed)
-	assert.Error(t, err, "should return error")
-}
-
 func TestContainsAny(t *testing.T) {
 	assert.True(t, ContainsAny("some words", "some"), "should contain element")
 	assert.True(t, ContainsAny("some words", "unknown", "some"), "should contain element")
 	assert.False(t, ContainsAny("some words", "unknown", "unknown 2"), "should not contain any element")
-}
-
-func TestHasParameter(t *testing.T) {
-	assert.True(t, hasParameter("a", "a"), "should contain parameter")
-	assert.True(t, hasParameter("a", "#a"), "should contain parameter")
-	assert.False(t, hasParameter("b", "a"), "should not contain parameter")
-	assert.False(t, hasParameter("b", "#a"), "should not contain parameter")
-
-	assert.True(t, hasParameter("a", "a&b"), "should contain parameter")
-	assert.True(t, hasParameter("b", "#a&b"), "should contain parameter")
-	assert.False(t, hasParameter("c", "a&b"), "should not contain parameter")
-	assert.False(t, hasParameter("c", "#a&b"), "should not contain parameter")
-
-	assert.True(t, hasParameter("a&b", "a&b"), "should contain parameter")
-	assert.True(t, hasParameter("a&b", "#a&b"), "should contain parameter")
-	assert.True(t, hasParameter("a&b", "a&b&c"), "should contain parameter")
-	assert.False(t, hasParameter("a&b&c", "a&b"), "should not contain parameter")
-
-	assert.True(t, hasParameter("b&a", "a&b"), "should contain parameter")
-	assert.True(t, hasParameter("b&a", "#a&b"), "should contain parameter")
-	assert.True(t, hasParameter("b&a", "a&b&c"), "should contain parameter")
-	assert.True(t, hasParameter("b&a", "#a&b&c"), "should contain parameter")
-
-	assert.False(t, hasParameter("a&c", "a&b"), "should not contain parameter")
-	assert.False(t, hasParameter("a&c", "#a&b"), "should not contain parameter")
 }
 
 func TestRemap(t *testing.T) {
